commit modificacion anagement
